backend/DetectifAI_db/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==3.0.0
PyJWT==2.8.0
pymongo>=4.6.3,<5.0
python-multipart==0.0.6
minio==7.1.11
opencv-python==4.8.0.74
python-dotenv==1.0.0
faiss-cpu
numpy
Pillow
scikit-learn
sentence-transformers
